move fixture helpers into fixture file
